+++ CHANGES.rst
@@ -11,6 +11,7 @@ Unreleased
 -   Deprecate :func:`utils.bind_arguments` and
     :func:`utils.validate_arguments`, use :meth:`Signature.bind` and
     :func:`inspect.signature` instead. :issue:`1757`
+-   Deprecate :class:`utils.HTMLBuilder`. :issue:`1761`
 -   ``JSONMixin`` no longer uses simplejson if it's installed. To use
     another JSON module, override ``JSONMixin.json_module``. :pr:`1766`
 -   Add a ``url_scheme`` argument to :meth:`~routing.MapAdapter.build`